refactor: convert to binary-only crate
The lib.rs was only there to support integration test imports,
which accidentally created a public API. Inline the app module
into the binary, reduce visibility to pub(crate) or private,
and move integration tests into the crate-internal test module.
